+++ gcovr/__main__.py
@@ -36,7 +36,7 @@ from .utils import (
     AlwaysMatchFilter,
     DirectoryPrefixFilter,
     configure_logging,
-    switch_to_logging_format_with_threads,
+    update_logging_formatter,
 )
 from .version import __version__
 from .coverage import CovData, SummarizedStats
@@ -195,9 +195,9 @@ def main(args=None):
                 parse_config_file(cfg_file, filename=cfg_name)
             )
     options = merge_options_and_set_defaults([cfg_options, cli_options.__dict__])
+
     # Reconfigure the logging.
-    if options.gcov_parallel > 1:
-        switch_to_logging_format_with_threads()
+    update_logging_formatter(options)
 
     if options.verbose:
         LOGGER.setLevel(logging.DEBUG)